last minute code update v3.7.0
git-svn-id: svn://svn.code.sf.net/p/steemsse/code/trunk/steemsse@341 dd3fbd3b-ef5d-4b78-98e8-da144fa02e5b

+++ steem/code/options.cpp
@@ -1959,8 +1959,8 @@ LRESULT __stdcall TOptionBox::WndProc(HWND Win,UINT Mess,WPARAM wPar,LPARAM lPar
             OPTION_CPU_CLOCK=!OPTION_CPU_CLOCK;
             TRACE_LOG("Option CPU clock = %d\n",OPTION_CPU_CLOCK);
             SendMessage(HWND(lPar),BM_SETCHECK,OPTION_CPU_CLOCK,0);
-            CpuMfpRatio=(OPTION_CPU_CLOCK?CpuCustomHz:CpuNormalHz)
-              /(double)MFP_CLK_TH_EXACT;
+            CpuMfpRatio=(OPTION_CPU_CLOCK?(double)CpuCustomHz
+              :(double)CpuNormalHz)  /(double)MFP_CLK_TH_EXACT;
           }
           break;
 #endif
